Compile on osx

+++ src/laser_calibration/CalibrationRectangle.cpp
@@ -21,7 +21,7 @@ std::vector<etherdream_point> CalibrationRectangle::points() const
     for (int i = 0; i < pointsPerLine; ++i)
     {
         cv::Point2f newPoint(m_topLeft.x + (i/pointsPerLine) * fromTo.x, m_topLeft.y + (i/pointsPerLine) * fromTo.y);
-        points.push_back(etherdream_point { newPoint.x, newPoint.y, 0, UINT16_MAX, 0 } );
+		points.push_back(etherdream_point { (int16_t) newPoint.x, (int16_t) newPoint.y, 0, UINT16_MAX, 0 } );
     }
 
     // bottomLeft - bottomRight
@@ -29,7 +29,7 @@ std::vector<etherdream_point> CalibrationRectangle::points() const
     for (int i = 0; i < pointsPerLine; ++i)
     {
         cv::Point2f newPoint(m_bottomLeft.x + (i/pointsPerLine) * fromTo.x, m_bottomLeft.y + (i/pointsPerLine) * fromTo.y);
-        points.push_back(etherdream_point { newPoint.x, newPoint.y, 0, UINT16_MAX, 0 } );
+		points.push_back(etherdream_point { (int16_t) newPoint.x, (int16_t) newPoint.y, 0, UINT16_MAX, 0 } );
     }
 
     // bottomRight - topRight
@@ -37,7 +37,7 @@ std::vector<etherdream_point> CalibrationRectangle::points() const
     for (int i = 0; i < pointsPerLine; ++i)
     {
         cv::Point2f newPoint(m_bottomRight.x + (i/pointsPerLine) * fromTo.x, m_bottomRight.y + (i/pointsPerLine) * fromTo.y);
-        points.push_back(etherdream_point { newPoint.x, newPoint.y, 0, UINT16_MAX, 0 } );
+		points.push_back(etherdream_point { (int16_t) newPoint.x, (int16_t) newPoint.y, 0, UINT16_MAX, 0 } );
     }
 
     // topRight - topLeft
@@ -45,7 +45,7 @@ std::vector<etherdream_point> CalibrationRectangle::points() const
     for (int i = 0; i < pointsPerLine; ++i)
     {
         cv::Point2f newPoint(m_topRight.x + (i/pointsPerLine) * fromTo.x, m_topRight.y + (i/pointsPerLine) * fromTo.y);
-        points.push_back(etherdream_point { newPoint.x, newPoint.y, 0, UINT16_MAX, 0 } );
+		points.push_back(etherdream_point { (int16_t) newPoint.x, (int16_t) newPoint.y, 0, UINT16_MAX, 0 } );
     }
 
     return points;
